feat: cooldown, dash and bomb, limit base atk charge,...

+++ Assets/_app/Scripts/Core/Interface/IMovable.cs
@@ -4,6 +4,6 @@ namespace STD.Core.Interface
 {
     public interface IMovable
     {
-        void Move(Vector2 input, int speed, float dt);
+        void Move(Vector2 input, float speed, float dt);
     }
 }
